+++ src/test/kotlin/lotto/domain/LottoResultTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.*
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
