+++ src/test/kotlin/lotto/domain/LottoStoreTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import lotto.domain.LottoStore
 import lotto.domain.LottoStore.Companion.LOTTO_COST
